PycharmProjects/pythonProject4/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
config==0.5.1
crispy-bootstrap5==0.7
distlib==0.3.7
Django==4.2.2
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
exceptiongroup==1.1.3
filelock==3.12.2
generic==1.1.2
generics==6.0.0
manage.py==0.2.10
Pillow==10.0.0
pipenv==2023.8.23
platformdirs==3.10.0
pycparser==2.21
pystache==0.6.4
pytz==2023.3.post1
sqlparse==0.4.4
static==1.1.1
tzdata==2023.3
virtualenv==20.24.3
